centralized all textfield types in a single file
