+++ src/incubator/TextField/FieldContext.ts
@@ -1,15 +1,6 @@
 import _ from 'lodash';
 import {createContext} from 'react';
-
-export type FieldContextType = {
-  value?: string;
-  isFocused: boolean;
-  hasValue: boolean;
-  isValid: boolean;
-  failingValidatorIndex?: number;
-  disabled: boolean;
-  validateField: () => void
-};
+import {FieldContextType} from './types';
 
 const FieldContext = createContext<FieldContextType>({
   isFocused: false,
