+++ src/incubator/TextField/Label.tsx
@@ -1,32 +1,11 @@
 import React, {useContext} from 'react';
-import {StyleSheet, TextStyle} from 'react-native';
+import {StyleSheet} from 'react-native';
 import {Colors} from '../../style';
-import Text, {TextProps} from '../../components/text';
-import {ColorType, ValidationMessagePosition} from './types';
+import Text from '../../components/text';
+import {LabelProps, ValidationMessagePosition} from './types';
 import {getColorByState} from './Presenter';
 import FieldContext from './FieldContext';
 
-export interface LabelProps {
-  /**
-   * Field label
-   */
-  label?: string;
-  /**
-   * Field label color. Either a string or color by state map ({default, focus, error, disabled})
-   */
-  labelColor?: ColorType;
-  /**
-   * Custom style for the field label
-   */
-  labelStyle?: TextStyle;
-  /**
-   * Pass extra props to the label Text element
-   */
-  labelProps?: TextProps;
-  validationMessagePosition?: ValidationMessagePosition;
-  floatingPlaceholder?: boolean;
-  testID?: string;
-}
 
 const Label = ({
   label,
