+++ src/incubator/TextField/ValidationMessage.tsx
@@ -1,27 +1,9 @@
 import React, {useContext} from 'react';
-import {TextStyle, StyleSheet} from 'react-native';
+import {StyleSheet} from 'react-native';
 import Text from '../../components/text';
 import FieldContext from './FieldContext';
 import {getRelevantValidationMessage} from './Presenter';
-import {FieldStateProps} from './useFieldState';
-
-export interface ValidationMessageProps {
-  /**
-   * Should support showing validation error message
-   */
-  enableErrors?: boolean;
-  /**
-   * The validation message to display when field is invalid (depends on validate)
-   */
-  validationMessage?: string | string[];
-  /**
-   * Custom style for the validation message
-   */
-  validationMessageStyle?: TextStyle;
-  retainSpace?: boolean;
-  validate?: FieldStateProps['validate'];
-  testID?: string;
-}
+import {ValidationMessageProps} from './types';
 
 const ValidationMessage = ({
   validationMessage,
