+++ src/incubator/TextField/index.tsx
@@ -5,113 +5,22 @@
  * 3. Passing typography preset that includes lineHeight usually cause alignment issues with
  * other elements (leading/trailing accessories). It usually best to set lineHeight with undefined
  */
-import React, {PropsWithChildren, ReactElement, useMemo} from 'react';
-import {ViewStyle, TextStyle, StyleProp} from 'react-native';
+import React, {useMemo} from 'react';
 import {omit} from 'lodash';
-import {
-  asBaseComponent,
-  forwardRef,
-  ForwardRefInjectedProps,
-  BaseComponentInjectedProps,
-  MarginModifiers,
-  PaddingModifiers,
-  TypographyModifiers,
-  ColorsModifiers
-} from '../../commons/new';
+import {asBaseComponent, forwardRef} from '../../commons/new';
 import View from '../../components/view';
 import {Colors} from '../../style';
 import {useMeasure} from '../../hooks';
-import {ValidationMessagePosition, Validator} from './types';
+import {TextFieldProps, InternalTextFieldProps, ValidationMessagePosition, FieldContextType} from './types';
 import {shouldHidePlaceholder} from './Presenter';
-import Input, {InputProps} from './Input';
-import ValidationMessage, {ValidationMessageProps} from './ValidationMessage';
-import Label, {LabelProps} from './Label';
-import FieldContext, {FieldContextType as _FieldContextType} from './FieldContext';
-import useFieldState /* , FieldStateProps */ from './useFieldState';
+import Input from './Input';
+import ValidationMessage from './ValidationMessage';
+import Label from './Label';
+import FieldContext from './FieldContext';
+import useFieldState from './useFieldState';
 import usePreset from './usePreset';
-import FloatingPlaceholder, {FloatingPlaceholderProps} from './FloatingPlaceholder';
-import CharCounter, {CharCounterProps} from './CharCounter';
-
-export type FieldContextType = _FieldContextType;
-export type TextFieldProps = MarginModifiers &
-  PaddingModifiers &
-  TypographyModifiers &
-  ColorsModifiers &
-  InputProps &
-  LabelProps &
-  Omit<FloatingPlaceholderProps, 'testID'> &
-  // We're declaring these props explicitly here for react-docgen (which can't read hooks)
-  // FieldStateProps &
-  ValidationMessageProps &
-  Omit<CharCounterProps, 'maxLength' | 'testID'> & {
-    /**
-     * Pass to render a leading element
-     */
-    leadingAccessory?: ReactElement;
-    /**
-     * Pass to render a trailing element
-     */
-    trailingAccessory?: ReactElement;
-    /**
-     * Pass to render a bottom element below the input
-     */
-    bottomAccessory?: ReactElement;
-    /**
-     * Pass to add floating placeholder support
-     */
-    floatingPlaceholder?: boolean;
-    /**
-     * Custom style for the floating placeholder
-     */
-    floatingPlaceholderStyle?: TextStyle;
-    /**
-     * A single or multiple validator. Can be a string (required, email) or custom function.
-     */
-    validate?: Validator | Validator[];
-    /**
-     * Should validate when the TextField mounts
-     */
-    validateOnStart?: boolean;
-    /**
-     * Should validate when the TextField value changes
-     */
-    validateOnChange?: boolean;
-    /**
-     * Should validate when losing focus of TextField
-     */
-    validateOnBlur?: boolean;
-    /**
-     * Callback for when field validity has changed
-     */
-    onChangeValidity?: (isValid: boolean) => void;
-    /**
-     * The position of the validation message (top/bottom)
-     */
-    validationMessagePosition?: ValidationMessagePosition;
-    /**
-     * Internal style for the field container
-     */
-    fieldStyle?: StyleProp<ViewStyle>;
-    /**
-     * Internal dynamic style callback for the field container
-     */
-    dynamicFieldStyle?: (context: FieldContextType, props: {preset: TextFieldProps['preset']}) => StyleProp<ViewStyle>;
-    /**
-     * Container style of the whole component
-     */
-    containerStyle?: ViewStyle;
-    /**
-     * Predefined preset to use for styling the field
-     */
-    preset?: 'default' | null | string;
-  };
-
-export type InternalTextFieldProps = PropsWithChildren<
-  TextFieldProps &
-    // Omit<FieldStateInjectedProps, keyof InputProps> &
-    BaseComponentInjectedProps &
-    ForwardRefInjectedProps
->;
+import FloatingPlaceholder from './FloatingPlaceholder';
+import CharCounter from './CharCounter';
 
 interface StaticMembers {
   validationMessagePositions: typeof ValidationMessagePosition;
@@ -243,7 +152,13 @@ const TextField = (props: InternalTextFieldProps) => {
             />
           )}
           {bottomAccessory}
-          {showCharCounter && <CharCounter maxLength={others.maxLength} charCounterStyle={charCounterStyle} testID={`${props.testID}.charCounter`}/>}
+          {showCharCounter && (
+            <CharCounter
+              maxLength={others.maxLength}
+              charCounterStyle={charCounterStyle}
+              testID={`${props.testID}.charCounter`}
+            />
+          )}
         </View>
       </View>
     </FieldContext.Provider>
@@ -253,4 +168,5 @@ const TextField = (props: InternalTextFieldProps) => {
 TextField.displayName = 'Incubator.TextField';
 TextField.validationMessagePositions = ValidationMessagePosition;
 
+export {TextFieldProps, FieldContextType};
 export default asBaseComponent<TextFieldProps, StaticMembers>(forwardRef(TextField as any));
