+++ src/incubator/TextField/usePreset.ts
@@ -1,7 +1,7 @@
-import {InternalTextFieldProps} from './index';
 import defaultPreset from './presets/default';
+import {InternalTextFieldProps} from './types';
 
-export default function usePreset({preset/*  = 'default' */, ...props}: InternalTextFieldProps) {
+export default function usePreset({preset /*  = 'default' */, ...props}: InternalTextFieldProps) {
   if (preset === 'default') {
     return {...defaultPreset, ...props, fieldStyle: [defaultPreset.fieldStyle, props.fieldStyle]};
   }
